Perform grad step in atomic variable in integrator
The `joint_logprob` in `aeppl` clones value variables which are the
output of a graph. As a result, when we perform a `aesara.grad` step on
the output of `joint_logprob` on said value variable we get a
`GradientDisconnectedError`.

Here we perform the grad step on an atomic placeholder for the position
and then replace this placeholder with the actual position's graph.