Update the Dual Averaging algorithm's API
During window adaptation we need to be able to re-initialize the step
size adaptation state from the current value of the step size. This is
currently impossible since the initial step size is embedded in the
`update` function via a closure.

We thus modify the dual averaging implementation and the step size
adaptation based on dual averaging so that the `mu` parameter (taken to
be log(10 * step_size) in step size adaptation) is passed to the `init`
function and subsequently carried along in the state.