Pass parameters as input to NUTS and HMC kernels
We currently have to build a new kernel every time the value of either
the step size, inverse mass matrix or length of trajectory changes. This
makes adaptation code harder to write and read.

In this commit we ask for the current value of these parameters at each
step. Any ensuing redundancy should be eliminated at compile time by aesara.